+++ src/plans/cache.jl
@@ -2,7 +2,7 @@
 # A Simple Cache for Objectives
 #
 @doc raw"""
-     SimpleCacheObjective{O<:AbstractManifoldGradientObjective{E,TC,TG}, P, T,C} <: AbstractManifoldGradientObjective{E,TC,TG}
+     SimpleCachedManifoldObjective{O<:AbstractManifoldGradientObjective{E,TC,TG}, P, T,C} <: AbstractManifoldGradientObjective{E,TC,TG}
 
 Provide a simple cache for an [`AbstractManifoldGradientObjective`](@ref) that is for a given point `p` this cache
 stores a point `p` and a gradient ``\operatorname{grad} f(p)`` in `X` as well as a cost value ``f(p)`` in `c`.
@@ -11,7 +11,7 @@ Both `X` and `c` are accompanied by booleans to keep track of their validity.
 
 # Constructor
 
-    SimpleCacheObjective(M::AbstractManifold, obj::AbstractManifoldGradientObjective; kwargs...)
+    SimpleCachedManifoldObjective(M::AbstractManifold, obj::AbstractManifoldGradientObjective; kwargs...)
 
 ## Keyword
 * `p` (`rand(M)`) – a point on the manifold to initialize the cache with
@@ -19,7 +19,7 @@ Both `X` and `c` are accompanied by booleans to keep track of their validity.
 * `c` (`get_cost(M, obj, p)` or `0.0`) – a value to store the cost function in `initialize`
 * `initialized` (`true`) – whether to initialize the cached `X` and `c` or not.
 """
-mutable struct SimpleCacheObjective{
+mutable struct SimpleCachedManifoldObjective{
     E<:AbstractEvaluationType,TC,TG,O<:AbstractManifoldGradientObjective{E,TC,TG},P,T,C
 } <: AbstractDecoratedManifoldObjective{E,O}
     objective::O
@@ -30,7 +30,7 @@ mutable struct SimpleCacheObjective{
     c_valid::Bool
 end
 
-function SimpleCacheObjective(
+function SimpleCachedManifoldObjective(
     M::AbstractManifold,
     obj::O;
     initialized=true,
@@ -39,12 +39,12 @@ function SimpleCacheObjective(
     c=initialized ? get_cost(M, obj, p) : 0.0,
 ) where {E<:AbstractEvaluationType,TC,TG,O<:AbstractManifoldGradientObjective{E,TC,TG}}
     q = copy(M, p)
-    return SimpleCacheObjective{E,TC,TG,O,typeof(q),typeof(X),typeof(c)}(
+    return SimpleCachedManifoldObjective{E,TC,TG,O,typeof(q),typeof(X),typeof(c)}(
         obj, q, X, initialized, c, initialized
     )
 end
 # Default implementations
-function get_cost(M::AbstractManifold, sco::SimpleCacheObjective, p)
+function get_cost(M::AbstractManifold, sco::SimpleCachedManifoldObjective, p)
     scop_neq_p = sco.p != p
     if scop_neq_p || !sco.c_valid
         # else evaluate cost, invalidate grad if p changed
@@ -56,8 +56,8 @@ function get_cost(M::AbstractManifold, sco::SimpleCacheObjective, p)
     end
     return sco.c
 end
-get_cost_function(sco::SimpleCacheObjective) = get_cost_function(sco.objective)
-function get_gradient(M::AbstractManifold, sco::SimpleCacheObjective, p)
+get_cost_function(sco::SimpleCachedManifoldObjective) = get_cost_function(sco.objective)
+function get_gradient(M::AbstractManifold, sco::SimpleCachedManifoldObjective, p)
     scop_neq_p = sco.p != p
     if scop_neq_p || !sco.X_valid
         get_gradient!(M, sco.X, sco.objective, p)
@@ -68,7 +68,7 @@ function get_gradient(M::AbstractManifold, sco::SimpleCacheObjective, p)
     end
     return copy(M, p, sco.X)
 end
-function get_gradient!(M::AbstractManifold, X, sco::SimpleCacheObjective, p)
+function get_gradient!(M::AbstractManifold, X, sco::SimpleCachedManifoldObjective, p)
     scop_neq_p = sco.p != p
     if scop_neq_p || !sco.X_valid
         get_gradient!(M, sco.X, sco.objective, p)
@@ -80,14 +80,18 @@ function get_gradient!(M::AbstractManifold, X, sco::SimpleCacheObjective, p)
     copyto!(M, X, sco.p, sco.X)
     return X
 end
-get_gradient_function(sco::SimpleCacheObjective) = get_gradient_function(sco.objective)
+function get_gradient_function(sco::SimpleCachedManifoldObjective)
+    return get_gradient_function(sco.objective)
+end
 
 #
 # CostGradImplementation
 #
 function get_cost(
     M::AbstractManifold,
-    sco::SimpleCacheObjective{AllocatingEvaluation,TC,TG,<:ManifoldCostGradientObjective},
+    sco::SimpleCachedManifoldObjective{
+        AllocatingEvaluation,TC,TG,<:ManifoldCostGradientObjective
+    },
     p,
 ) where {TC,TG}
     scop_neq_p = sco.p != p
@@ -101,7 +105,9 @@ function get_cost(
 end
 function get_cost(
     M::AbstractManifold,
-    sco::SimpleCacheObjective{InplaceEvaluation,TC,TG,<:ManifoldCostGradientObjective},
+    sco::SimpleCachedManifoldObjective{
+        InplaceEvaluation,TC,TG,<:ManifoldCostGradientObjective
+    },
     p,
 ) where {TC,TG}
     scop_neq_p = sco.p != p
@@ -115,7 +121,9 @@ function get_cost(
 end
 function get_gradient(
     M::AbstractManifold,
-    sco::SimpleCacheObjective{AllocatingEvaluation,TC,TG,<:ManifoldCostGradientObjective},
+    sco::SimpleCachedManifoldObjective{
+        AllocatingEvaluation,TC,TG,<:ManifoldCostGradientObjective
+    },
     p,
 ) where {TC,TG}
     scop_neq_p = sco.p != p
@@ -130,7 +138,9 @@ function get_gradient(
 end
 function get_gradient(
     M::AbstractManifold,
-    sco::SimpleCacheObjective{InplaceEvaluation,TC,TG,<:ManifoldCostGradientObjective},
+    sco::SimpleCachedManifoldObjective{
+        InplaceEvaluation,TC,TG,<:ManifoldCostGradientObjective
+    },
     p,
 ) where {TC,TG}
     scop_neq_p = sco.p != p
@@ -146,7 +156,9 @@ end
 function get_gradient!(
     M::AbstractManifold,
     X,
-    sco::SimpleCacheObjective{AllocatingEvaluation,TC,TG,<:ManifoldCostGradientObjective},
+    sco::SimpleCachedManifoldObjective{
+        AllocatingEvaluation,TC,TG,<:ManifoldCostGradientObjective
+    },
     p,
 ) where {TC,TG}
     scop_neq_p = sco.p != p
@@ -162,7 +174,9 @@ end
 function get_gradient!(
     M::AbstractManifold,
     X,
-    sco::SimpleCacheObjective{InplaceEvaluation,TC,TG,<:ManifoldCostGradientObjective},
+    sco::SimpleCachedManifoldObjective{
+        InplaceEvaluation,TC,TG,<:ManifoldCostGradientObjective
+    },
     p,
 ) where {TC,TG}
     scop_neq_p = sco.p != p
@@ -177,7 +191,165 @@ function get_gradient!(
 end
 
 #
+# CachedManifoldObjective constructor which errors by default since we can only define init
+# for LRU Caches
+#
+@doc raw"""
+    CachedManifoldObjective{E,P,O<:AbstractManifoldObjective{<:E},C<:NamedTuple{}} <: AbstractDecoratedManifoldObjective{E,P}
+
+Create a cache for an objective, based on a `NamedTuple` that stores `LRUCaches` for
+
+# Constructor
+
+    CachedManifoldObjective(M, o::AbstractManifoldObjective, caches::Vector{Symbol}; kwargs...)
+
+Create a cache for the [`AbstractManifoldObjective`](@ref) where the Symbols in `caches` indicate,
+which function evaluations to cache.
+
+# Keyword Arguments
+* `p`           - (`rand(M)`) the type of the keys to be used in the caches. Defaults to the default representation on `M`.
+* `v`           - (`get_cost(M, objective, p)`) the type of values for numeric values in the cache, e.g. the cost
+* `X`           - (`zero_vector(M,p)`) the type of values to be cached for gradient and Hessian calls.
+* `cache`       - (`[:Cost]`) a vector of symbols indicating which function calls should be cached.
+* `cache_size`  - (`10`) number of (least recently used) calls to cache
+* `cache_sizes` – (`Dict{Symbol,Int}()`) a named tuple or dictionary specifying the sizes individually for each cache.
+"""
+struct CachedManifoldObjective{E,P,O<:AbstractManifoldObjective{<:E},C<:NamedTuple{}} <:
+       AbstractDecoratedManifoldObjective{E,P}
+    objective::O
+    cache::C
+end
+function CachedManifoldObjective(
+    M::AbstractManifold,
+    objective::O,
+    caches::AbstractVector{<:Symbol}=[:Cost];
+    p::P=rand(M),
+    v::R=get_cost(M, objective, p),
+    X::T=zero_vector(M, p),
+    cache_size=10,
+    cache_sizes=Dict{Symbol,Int}(),
+) where {E,O<:AbstractManifoldObjective{E},R,P,T}
+    c = init_caches(
+        M, caches; p=p, v=v, X=X, cache_size=cache_size, cache_sizes=cache_sizes
+    )
+    return CachedManifoldObjective{E,O,O,typeof(c)}(objective, c)
+end
+function CachedManifoldObjective(
+    M::AbstractManifold,
+    objective::O,
+    caches::AbstractVector{<:Symbol}=[:Cost];
+    p::P=rand(M),
+    v::R=get_cost(M, objective, p),
+    X::T=zero_vector(M, p),
+    cache_size=10,
+    cache_sizes=Dict{Symbol,Int}(),
+) where {E,O2,O<:AbstractDecoratedManifoldObjective{E,O2},R,P,T}
+    c = init_caches(
+        M, caches; p=p, v=v, X=X, cache_size=cache_size, cache_sizes=cache_sizes
+    )
+    return CachedManifoldObjective{E,O2,O,typeof(c)}(objective, c)
+end
+
+"""
+    init_caches(M::AbstractManifold, caches, T; kwargs...)
+
+Given a vector of symbols `caches`, this function sets up the
+`NamedTuple` of caches for points/vectors on `M`,
+where `T` is the type of cache to use.
+"""
+function init_caches(M, caches, T=Nothing; kwargs...)
+    return throw(DomainError(
+        T,
+        """
+        No function `init_caches` available for a caches $T.
+        For a good default load `LRUCache.jl`, for example:
+        Enter `using LRUCache`.
+        """,
+    ))
+end
+
+get_gradient_function(co::CachedManifoldObjective) = get_gradient_function(co.objective)
+get_cost_function(co::CachedManifoldObjective) = get_cost_function(co.objective)
+
+#
+# Default implementations - (a) check if field exists (b) try to get cache
+function get_cost(M::AbstractManifold, co::CachedManifoldObjective, p)
+    !(haskey(co.cache, :Cost)) && return get_cost(M, co.objective, p)
+    return get!(co.cache[:Cost], p) do
+        get_cost(M, co.objective, p)
+    end
+end
+
+function get_gradient(M::AbstractManifold, co::CachedManifoldObjective, p)
+    !(haskey(co.cache, :Gradient)) && return get_gradient(M, co.objective, p)
+    return get!(co.cache[:Gradient], p) do
+        get_gradient(M, co.objective, p)
+    end
+end
+function get_gradient!(M::AbstractManifold, X, co::CachedManifoldObjective, p)
+    !(haskey(co.cache, :Gradient)) && return get_gradient!(M, X, co.objective, p)
+    copyto!(
+        M,
+        X,
+        p,
+        get!(co.cache[:Gradient], p) do
+            get_gradient!(M, X, co.objective, p)
+        end,
+    )
+    return X
+end
+
+#
+# CostGradImplementation
+function get_cost(
+    M::AbstractManifold, co::CachedManifoldObjective{E,<:ManifoldCostGradientObjective}, p
+) where {E<:AbstractEvaluationType}
+    #Neither cost not grad cached -> evaluate
+    all(.!(haskey.(Ref(co.cache), [:Cost, :Gradient]))) &&
+        return get_cost(M, co.objective, p)
+    return get!(co.cache[:Cost], p) do
+        c, X = get_cost_and_gradient(M, co.objective, p)
+        #if this is evaluated, we can also set X
+        haskey(co.cache, :Gradient) && setindex!(co.cache[:Gradient], X, p)
+        c #but we also set the new cost here
+    end
+end
+function get_gradient(
+    M::AbstractManifold, co::CachedManifoldObjective{E,<:ManifoldCostGradientObjective}, p
+) where {E<:AllocatingEvaluation}
+    all(.!(haskey.(Ref(co.cache), [:Cost, :Gradient]))) &&
+        return get_gradient(M, co.objective, p)
+    return get!(co.cache[:Gradient], p) do
+        c, X = get_cost_and_gradient(M, co.objective, p)
+        #if this is evaluated, we can also set c
+        haskey(co.cache, :Cost) && setindex!(co.cache[:Cost], c, p)
+        X #but we also set the new cost here
+    end
+end
+function get_gradient!(
+    M::AbstractManifold,
+    X,
+    co::CachedManifoldObjective{E,<:ManifoldCostGradientObjective},
+    p,
+) where {E}
+    All(!(haskey.(Ref(co.cache), [:Cost, :Gradient]))) &&
+        return get_gradient!(M, X, co.objective, p)
+    return copyto!(
+        M,
+        X,
+        p,
+        get!(co.cache[:Gradient], p) do
+            c, _ = get_cost_and_gradient!(M, X, co.objective, p)
+            #if this is evaluated, we can also set c
+            haskey(co.cache, :Cost) && setindex!(co.cache[:Cost], c, p)
+            X
+        end,
+    )
+end
+
+#
 # Factory
+#
 @doc raw"""
     objective_cache_factory(M::AbstractManifold, o::AbstractManifoldObjective, cache::Symbol)
 
@@ -186,16 +358,19 @@ on the `AbstractManifold M` based on the symbol `cache`.
 
 The following caches are available
 
-* `:Simple` generates a [`SimpleCacheObjective`](@ref)
-* `:LRU` generates a [`LRUCacheObjective`](@ref) where you should use the form
+* `:Simple` generates a [`SimpleCachedManifoldObjective`](@ref)
+* `:LRU` generates a [`CachedManifoldObjective`](@ref) where you should use the form
   `(:LRU, [:Cost, :Gradient])` to specify what should be cached or
   `(:LRU, [:Cost, :Gradient], 100)` to specify the cache size.
   Here this variant defaults to `(:LRU, [:Cost, :Gradient], 100)`,
-  i.e. to cache up to 100 cost and gradient values.
+  i.e. to cache up to 100 cost and gradient values.[^1]
+
+[^1] this cache requires [`LRUCache.jl`](https://github.com/JuliaCollections/LRUCache.jl) to be loaded as well.
 """
 function objective_cache_factory(M, o, cache::Symbol)
-    (cache === :Simple) && return SimpleCacheObjective(M, o)
-    (cache === :LRU) && return LRUCacheObjective(M, o, [:Cost, :Gradient]; cache_size=100)
+    (cache === :Simple) && return SimpleCachedManifoldObjective(M, o)
+    (cache === :LRU) &&
+        return CachedManifoldObjective(M, o, [:Cost, :Gradient]; cache_size=100)
     return o
 end
 
@@ -211,18 +386,18 @@ the optional third is passed down as keyword arguments.
 For all available caches see the simpler variant with symbols.
 """
 function objective_cache_factory(M, o, cache::Tuple{Symbol,<:AbstractArray,<:AbstractArray})
-    (cache[1] === :Simple) && return SimpleCacheObjective(M, o; cache[3]...)
+    (cache[1] === :Simple) && return SimpleCachedManifoldObjective(M, o; cache[3]...)
     if (cache[1] === :LRU)
         if (cacge[3] isa Integer)
-            return LRUCacheObjective(M, o, cache[2]; cache_size=cache[3])
+            return CachedManifoldObjective(M, o, cache[2]; cache_size=cache[3])
         else
-            return LRUCacheObjective(M, o, cache[2]; cache[3]...)
+            return CachedManifoldObjective(M, o, cache[2]; cache[3]...)
         end
     end
     return o
 end
 function objective_cache_factory(M, o, cache::Tuple{Symbol,<:AbstractArray})
-    (cache[1] === :Simple) && return SimpleCacheObjective(M, o)
+    (cache[1] === :Simple) && return SimpleCachedManifoldObjective(M, o)
     (cache[1] === :LRU) && return objective_cache_factory(M, o, Val(:LRU), cache[2])
     return o
 end